Better separate simple-tree and shared-tree
