+++ packages/dds/tree/src/index.ts
@@ -61,10 +61,7 @@ export {
 	type SharedTreeFormatOptions,
 	SharedTreeFormatVersion,
 	Tree,
-	type TransactionConstraint,
-	type NodeInDocumentConstraint,
 	type RunTransaction,
-	rollback,
 	type ForestOptions,
 	getBranch,
 	type BranchableTree,
@@ -73,18 +70,9 @@ export {
 	type ViewContent,
 	TreeAlpha,
 	independentView,
-	type RunTransactionParams,
-	type VoidTransactionCallbackStatus,
-	type TransactionCallbackStatus,
-	type TransactionResult,
-	type TransactionResultExt,
-	type TransactionResultSuccess,
-	type TransactionResultFailed,
 	ForestTypeOptimized,
 	ForestTypeExpensiveDebug,
 	ForestTypeReference,
-	SharedTreeAttributes,
-	SharedTreeFactoryType,
 	type IChannelView,
 } from "./shared-tree/index.js";
 
@@ -218,10 +206,22 @@ export {
 	type NodeSchemaOptions,
 	type NodeSchemaMetadata,
 	type ITreeAlpha,
+	type TransactionConstraint,
+	type NodeInDocumentConstraint,
+	type RunTransactionParams,
+	type VoidTransactionCallbackStatus,
+	type TransactionCallbackStatus,
+	type TransactionResult,
+	type TransactionResultExt,
+	type TransactionResultSuccess,
+	type TransactionResultFailed,
+	rollback,
 } from "./simple-tree/index.js";
 export {
 	SharedTree,
 	configuredSharedTree,
+	SharedTreeAttributes,
+	SharedTreeFactoryType,
 } from "./treeFactory.js";
 
 export {
