+++ packages/dds/tree/src/shared-tree/index.ts
@@ -25,11 +25,6 @@ export {
 } from "./sharedTree.js";
 
 export {
-	SharedTreeAttributes,
-	SharedTreeFactoryType,
-} from "./publicContracts.js";
-
-export {
 	createTreeCheckout,
 	TreeCheckout,
 	type ITreeCheckout,
@@ -52,19 +47,6 @@ export {
 	type TreeApi,
 	type RunTransaction,
 } from "./treeApi.js";
-
-export {
-	type TransactionConstraint,
-	type NodeInDocumentConstraint,
-	type RunTransactionParams,
-	type VoidTransactionCallbackStatus,
-	type TransactionCallbackStatus,
-	type TransactionResult,
-	type TransactionResultExt,
-	type TransactionResultSuccess,
-	type TransactionResultFailed,
-	rollback,
-} from "./transactionTypes.js";
 
 export { TreeAlpha } from "./treeApiAlpha.js";
