+++ packages/dds/tree/src/simple-tree/index.ts
@@ -125,6 +125,16 @@ export {
 	type CustomTreeValue,
 	tryStoredSchemaAsArray,
 	type ITreeAlpha,
+	type TransactionConstraint,
+	type NodeInDocumentConstraint,
+	type RunTransactionParams,
+	type VoidTransactionCallbackStatus,
+	type TransactionCallbackStatus,
+	type TransactionResult,
+	type TransactionResultExt,
+	type TransactionResultSuccess,
+	type TransactionResultFailed,
+	rollback,
 } from "./api/index.js";
 export {
 	type NodeFromSchema,
