+++ packages/dds/tree/src/test/shared-tree/treeNodeApi.spec.ts
@@ -7,13 +7,7 @@ import { strict as assert } from "node:assert";
 
 import { MockHandle } from "@fluidframework/test-runtime-utils/internal";
 
-import {
-	CheckoutFlexTreeView,
-	type TransactionConstraint,
-	Tree,
-	TreeAlpha,
-	type rollback,
-} from "../../shared-tree/index.js";
+import { CheckoutFlexTreeView, Tree, TreeAlpha } from "../../shared-tree/index.js";
 import {
 	SchemaFactory,
 	TreeViewConfiguration,
@@ -24,6 +18,8 @@ import {
 	type NodeFromSchema,
 	asTreeViewAlpha,
 	type TreeViewAlpha,
+	type TransactionConstraint,
+	type rollback,
 } from "../../simple-tree/index.js";
 import {
 	TestTreeProviderLite,
