+++ packages/dds/tree/src/treeFactory.ts
@@ -16,13 +16,31 @@ import {
 } from "@fluidframework/shared-object-base/internal";
 
 import {
-	SharedTreeAttributes,
-	SharedTreeFactoryType,
 	SharedTree as SharedTreeImpl,
 	type SharedTreeOptions,
 	type SharedTreeOptionsInternal,
 } from "./shared-tree/index.js";
 import type { ITree } from "./simple-tree/index.js";
+
+import { pkgVersion } from "./packageVersion.js";
+
+/**
+ * {@inheritDoc @fluidframework/shared-object-base#ISharedObjectFactory."type"}
+ * @alpha
+ * @legacy
+ */
+export const SharedTreeFactoryType = "https://graph.microsoft.com/types/tree";
+
+/**
+ * {@inheritDoc @fluidframework/shared-object-base#ISharedObjectFactory.attributes}
+ * @alpha
+ * @legacy
+ */
+export const SharedTreeAttributes: IChannelAttributes = {
+	type: SharedTreeFactoryType,
+	snapshotFormatVersion: "0.0.0",
+	packageVersion: pkgVersion,
+};
 
 /**
  * A channel factory that creates an {@link ITree}.
